pygui/requirements.txt
gi
pyserial
